moves facilitator types to its own file
